Set amount to be double instead of int

+++ app/src/main/java/me/anon/model/Feed.java
@@ -15,5 +15,5 @@ import lombok.experimental.Accessors;
 public class Feed extends Water
 {
 	private Nutrient nutrient;
-	private Integer mlpl; // ml per litre
+	private Double mlpl; // ml per litre
 }
